+++ etils/ecolab/array_as_img_test.py
@@ -57,7 +57,7 @@ def test_array_repr_html_valid(
     valid_shape: tuple[int, ...],
 ):
   # 2D images are displayed as images
-  assert '<img' in array_as_img._array_repr_html(
+  assert '<img' in array_as_img.array_repr_html(
       xnp.zeros(valid_shape), **_ARRAY_REPR_HTML_KWARGS
   )
 
@@ -74,7 +74,7 @@ def test_array_repr_video_html_valid(
     valid_shape: tuple[int, ...],
 ):
   # 2D images are displayed as video
-  assert '<video' in array_as_img._array_repr_html(
+  assert '<video' in array_as_img.array_repr_html(
       xnp.zeros(valid_shape), **_ARRAY_REPR_HTML_KWARGS
   )
 
@@ -99,7 +99,7 @@ def test_array_repr_html_invalid(
     invalid_shape: tuple[int, ...],
 ):
   assert (
-      array_as_img._array_repr_html(
+      array_as_img.array_repr_html(
           xnp.zeros(invalid_shape), **_ARRAY_REPR_HTML_KWARGS
       )
       is None